+++ app/Http/Controllers/Accessories/AccessoriesController.php
@@ -1,7 +1,8 @@
 <?php
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Accessories;
 
 use App\Helpers\Helper;
+use App\Http\Controllers\Controller;
 use App\Http\Requests\ImageUploadRequest;
 use App\Models\Accessory;
 use App\Models\Company;
@@ -197,161 +198,4 @@ class AccessoriesController extends Controller
         }
         return redirect()->route('accessories.index')->with('error', trans('admin/accessories/message.does_not_exist', compact('id')));
     }
-
-  /**
-   * Return the form to checkout an Accessory to a user.
-   *
-   * @author [A. Gianotto] [<snipe@snipe.net>]
-   * @param  int  $accessoryId
-   * @return View
-   */
-    public function getCheckout(Request $request, $accessoryId)
-    {
-        // Check if the accessory exists
-        if (is_null($accessory = Accessory::find($accessoryId))) {
-            // Redirect to the accessory management page with error
-            return redirect()->route('accessories.index')->with('error', trans('admin/accessories/message.not_found'));
-        }
-
-        if ($accessory->category) {
-
-            $this->authorize('checkout', $accessory);
-
-            // Get the dropdown of users and then pass it to the checkout view
-            return view('accessories/checkout', compact('accessory'));
-        }
-
-        return redirect()->back()->with('error', 'The category type for this accessory is not valid. Edit the accessory and select a valid accessory category.');
-
-
-
-    }
-
-  /**
-   * Save the Accessory checkout information.
-   *
-   * If Slack is enabled and/or asset acceptance is enabled, it will also
-   * trigger a Slack message and send an email.
-   *
-   * @author [A. Gianotto] [<snipe@snipe.net>]
-   * @param  int  $accessoryId
-   * @return Redirect
-   */
-    public function postCheckout(Request $request, $accessoryId)
-    {
-      // Check if the accessory exists
-        if (is_null($accessory = Accessory::find($accessoryId))) {
-            // Redirect to the accessory management page with error
-            return redirect()->route('accessories.index')->with('error', trans('admin/accessories/message.user_not_found'));
-        }
-
-        $this->authorize('checkout', $accessory);
-
-        if (!$user = User::find(Input::get('assigned_to'))) {
-            return redirect()->route('checkout/accessory', $accessory->id)->with('error', trans('admin/accessories/message.checkout.user_does_not_exist'));
-        }
-
-      // Update the accessory data
-        $accessory->assigned_to = e(Input::get('assigned_to'));
-
-        $accessory->users()->attach($accessory->id, [
-            'accessory_id' => $accessory->id,
-            'created_at' => Carbon::now(),
-            'user_id' => Auth::id(),
-            'assigned_to' => $request->get('assigned_to')
-        ]);
-
-        $logaction = $accessory->logCheckout(e(Input::get('note')), $user);
-
-        DB::table('accessories_users')->where('assigned_to', '=', $accessory->assigned_to)->where('accessory_id', '=', $accessory->id)->first();
-
-        $data['log_id'] = $logaction->id;
-        $data['eula'] = $accessory->getEula();
-        $data['first_name'] = $user->first_name;
-        $data['item_name'] = $accessory->name;
-        $data['checkout_date'] = $logaction->created_at;
-        $data['item_tag'] = '';
-        $data['expected_checkin'] = '';
-        $data['note'] = $logaction->note;
-        $data['require_acceptance'] = $accessory->requireAcceptance();
-
-      // Redirect to the new accessory page
-        return redirect()->route('accessories.index')->with('success', trans('admin/accessories/message.checkout.success'));
-    }
-
-
-    /**
-     * Check the accessory back into inventory
-     *
-     * @author [A. Gianotto] [<snipe@snipe.net>]
-     * @param Request $request
-     * @param integer $accessoryUserId
-     * @param string $backto
-     * @return View
-     * @internal param int $accessoryId
-     */
-    public function getCheckin(Request $request, $accessoryUserId = null, $backto = null)
-    {
-        // Check if the accessory exists
-        if (is_null($accessory_user = DB::table('accessories_users')->find($accessoryUserId))) {
-            // Redirect to the accessory management page with error
-            return redirect()->route('accessories.index')->with('error', trans('admin/accessories/message.not_found'));
-        }
-
-        $accessory = Accessory::find($accessory_user->accessory_id);
-        $this->authorize('checkin', $accessory);
-        return view('accessories/checkin', compact('accessory'))->with('backto', $backto);
-    }
-
-
-    /**
-     * Check in the item so that it can be checked out again to someone else
-     *
-     * @uses Accessory::checkin_email() to determine if an email can and should be sent
-     * @author [A. Gianotto] [<snipe@snipe.net>]
-     * @param Request $request
-     * @param integer $accessoryUserId
-     * @param string $backto
-     * @return Redirect
-     * @internal param int $accessoryId
-     */
-    public function postCheckin(Request $request, $accessoryUserId = null, $backto = null)
-    {
-      // Check if the accessory exists
-        if (is_null($accessory_user = DB::table('accessories_users')->find($accessoryUserId))) {
-            // Redirect to the accessory management page with error
-            return redirect()->route('accessories.index')->with('error', trans('admin/accessories/message.does_not_exist'));
-        }
-
-        $accessory = Accessory::find($accessory_user->accessory_id);
-
-        $this->authorize('checkin', $accessory);
-
-        $return_to = e($accessory_user->assigned_to);
-        $logaction = $accessory->logCheckin(User::find($return_to), e(Input::get('note')));
-
-        // Was the accessory updated?
-        if (DB::table('accessories_users')->where('id', '=', $accessory_user->id)->delete()) {
-            if (!is_null($accessory_user->assigned_to)) {
-                $user = User::find($accessory_user->assigned_to);
-            }
-
-            $data['log_id'] = $logaction->id;
-            $data['first_name'] = e($user->first_name);
-            $data['last_name'] = e($user->last_name);
-            $data['item_name'] = e($accessory->name);
-            $data['checkin_date'] = e($logaction->created_at);
-            $data['item_tag'] = '';
-            $data['note'] = e($logaction->note);
-
-            if ($backto=='user') {
-                return redirect()->route("users.show", $return_to)->with('success', trans('admin/accessories/message.checkin.success'));
-            }
-            return redirect()->route("accessories.show", $accessory->id)->with('success', trans('admin/accessories/message.checkin.success'));
-        }
-        // Redirect to the accessory management page with error
-        return redirect()->route('accessories.index')->with('error', trans('admin/accessories/message.checkin.error'));
-    }
-
-
 }
