+++ app/Http/Controllers/Accessories/AccessoryCheckinController.php
@@ -1,0 +1,85 @@
+<?php
+
+namespace App\Http\Controllers\Accessories;
+
+use App\Http\Controllers\Controller;
+use App\Models\Accessory;
+use App\Models\User;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Input;
+
+class AccessoryCheckinController extends Controller
+{
+        /**
+     * Check the accessory back into inventory
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @param Request $request
+     * @param integer $accessoryUserId
+     * @param string $backto
+     * @return View
+     * @internal param int $accessoryId
+     */
+    public function create(Request $request, $accessoryUserId = null, $backto = null)
+    {
+        // Check if the accessory exists
+        if (is_null($accessory_user = DB::table('accessories_users')->find($accessoryUserId))) {
+            // Redirect to the accessory management page with error
+            return redirect()->route('accessories.index')->with('error', trans('admin/accessories/message.not_found'));
+        }
+
+        $accessory = Accessory::find($accessory_user->accessory_id);
+        $this->authorize('checkin', $accessory);
+        return view('accessories/checkin', compact('accessory'))->with('backto', $backto);
+    }
+
+    /**
+     * Check in the item so that it can be checked out again to someone else
+     *
+     * @uses Accessory::checkin_email() to determine if an email can and should be sent
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @param Request $request
+     * @param integer $accessoryUserId
+     * @param string $backto
+     * @return Redirect
+     * @internal param int $accessoryId
+     */
+    public function store(Request $request, $accessoryUserId = null, $backto = null)
+    {
+      // Check if the accessory exists
+        if (is_null($accessory_user = DB::table('accessories_users')->find($accessoryUserId))) {
+            // Redirect to the accessory management page with error
+            return redirect()->route('accessories.index')->with('error', trans('admin/accessories/message.does_not_exist'));
+        }
+
+        $accessory = Accessory::find($accessory_user->accessory_id);
+
+        $this->authorize('checkin', $accessory);
+
+        $return_to = e($accessory_user->assigned_to);
+        $logaction = $accessory->logCheckin(User::find($return_to), e(Input::get('note')));
+
+        // Was the accessory updated?
+        if (DB::table('accessories_users')->where('id', '=', $accessory_user->id)->delete()) {
+            if (!is_null($accessory_user->assigned_to)) {
+                $user = User::find($accessory_user->assigned_to);
+            }
+
+            $data['log_id'] = $logaction->id;
+            $data['first_name'] = e($user->first_name);
+            $data['last_name'] = e($user->last_name);
+            $data['item_name'] = e($accessory->name);
+            $data['checkin_date'] = e($logaction->created_at);
+            $data['item_tag'] = '';
+            $data['note'] = e($logaction->note);
+
+            if ($backto=='user') {
+                return redirect()->route("users.show", $return_to)->with('success', trans('admin/accessories/message.checkin.success'));
+            }
+            return redirect()->route("accessories.show", $accessory->id)->with('success', trans('admin/accessories/message.checkin.success'));
+        }
+        // Redirect to the accessory management page with error
+        return redirect()->route('accessories.index')->with('error', trans('admin/accessories/message.checkin.error'));
+    }
+}
