+++ app/Http/Controllers/Accessories/AccessoryCheckoutController.php
@@ -1,0 +1,94 @@
+<?php
+
+namespace App\Http\Controllers\Accessories;
+
+use App\Http\Controllers\Controller;
+use App\Models\Accessory;
+use App\Models\User;
+use Carbon\Carbon;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Input;
+
+class AccessoryCheckoutController extends Controller
+{
+
+    /**
+   * Return the form to checkout an Accessory to a user.
+   *
+   * @author [A. Gianotto] [<snipe@snipe.net>]
+   * @param  int  $accessoryId
+   * @return View
+   */
+    public function create(Request $request, $accessoryId)
+    {
+        // Check if the accessory exists
+        if (is_null($accessory = Accessory::find($accessoryId))) {
+            // Redirect to the accessory management page with error
+            return redirect()->route('accessories.index')->with('error', trans('admin/accessories/message.not_found'));
+        }
+
+        if ($accessory->category) {
+
+            $this->authorize('checkout', $accessory);
+
+            // Get the dropdown of users and then pass it to the checkout view
+            return view('accessories/checkout', compact('accessory'));
+        }
+
+        return redirect()->back()->with('error', 'The category type for this accessory is not valid. Edit the accessory and select a valid accessory category.');
+    }
+
+     /**
+   * Save the Accessory checkout information.
+   *
+   * If Slack is enabled and/or asset acceptance is enabled, it will also
+   * trigger a Slack message and send an email.
+   *
+   * @author [A. Gianotto] [<snipe@snipe.net>]
+   * @param  int  $accessoryId
+   * @return Redirect
+   */
+    public function store(Request $request, $accessoryId)
+    {
+      // Check if the accessory exists
+        if (is_null($accessory = Accessory::find($accessoryId))) {
+            // Redirect to the accessory management page with error
+            return redirect()->route('accessories.index')->with('error', trans('admin/accessories/message.user_not_found'));
+        }
+
+        $this->authorize('checkout', $accessory);
+
+        if (!$user = User::find(Input::get('assigned_to'))) {
+            return redirect()->route('checkout/accessory', $accessory->id)->with('error', trans('admin/accessories/message.checkout.user_does_not_exist'));
+        }
+
+      // Update the accessory data
+        $accessory->assigned_to = e(Input::get('assigned_to'));
+
+        $accessory->users()->attach($accessory->id, [
+            'accessory_id' => $accessory->id,
+            'created_at' => Carbon::now(),
+            'user_id' => Auth::id(),
+            'assigned_to' => $request->get('assigned_to')
+        ]);
+
+        $logaction = $accessory->logCheckout(e(Input::get('note')), $user);
+
+        DB::table('accessories_users')->where('assigned_to', '=', $accessory->assigned_to)->where('accessory_id', '=', $accessory->id)->first();
+
+        $data['log_id'] = $logaction->id;
+        $data['eula'] = $accessory->getEula();
+        $data['first_name'] = $user->first_name;
+        $data['item_name'] = $accessory->name;
+        $data['checkout_date'] = $logaction->created_at;
+        $data['item_tag'] = '';
+        $data['expected_checkin'] = '';
+        $data['note'] = $logaction->note;
+        $data['require_acceptance'] = $accessory->requireAcceptance();
+
+      // Redirect to the new accessory page
+        return redirect()->route('accessories.index')->with('success', trans('admin/accessories/message.checkout.success'));
+    }
+}
