+++ app/Http/Controllers/Assets/AssetCheckoutController.php
@@ -1,9 +1,11 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Assets;
+
 
 use App\Exceptions\CheckoutNotAllowed;
 use App\Http\Controllers\CheckInOutRequest;
+use App\Http\Controllers\Controller;
 use App\Http\Requests\AssetCheckoutRequest;
 use App\Models\Asset;
 use App\Models\Location;
