+++ app/Http/Controllers/CompaniesController.php
@@ -2,10 +2,6 @@
 namespace App\Http\Controllers;
 
 use App\Models\Company;
-use Input;
-use Lang;
-use Redirect;
-use View;
 use Illuminate\Http\Request;
 use Image;
 use App\Http\Requests\ImageUploadRequest;
@@ -21,11 +17,12 @@ final class CompaniesController extends Controller
 {
 
     /**
-    * Returns view to display listing of companies.
-    *
-    * @author [Abdullah Alansari] [<ahimta@gmail.com>]
-    * @since [v1.8]
-    * @return \Illuminate\Contracts\View\View
+     * Returns view to display listing of companies.
+     *
+     * @author [Abdullah Alansari] [<ahimta@gmail.com>]
+     * @since [v1.8]
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index()
     {
@@ -35,11 +32,12 @@ final class CompaniesController extends Controller
     }
 
     /**
-    * Returns view to create a new company.
-    *
-    * @author [Abdullah Alansari] [<ahimta@gmail.com>]
-    * @since [v1.8]
-    * @return \Illuminate\Contracts\View\View
+     * Returns view to create a new company.
+     *
+     * @author [Abdullah Alansari] [<ahimta@gmail.com>]
+     * @since [v1.8]
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
@@ -55,6 +53,7 @@ final class CompaniesController extends Controller
      * @since [v1.8]
      * @param Request $request
      * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(ImageUploadRequest $request)
     {
@@ -63,16 +62,7 @@ final class CompaniesController extends Controller
         $company = new Company;
         $company->name = $request->input('name');
 
-        if ($request->file('image')) {
-            $image = $request->file('image');
-            $file_name = str_random(25).".".$image->getClientOriginalExtension();
-            $path = public_path('uploads/companies/'.$file_name);
-            Image::make($image->getRealPath())->resize(200, null, function ($constraint) {
-                $constraint->aspectRatio();
-                $constraint->upsize();
-            })->save($path);
-            $company->image = $file_name;
-        }
+        $company = $request->handleImages($company);
 
         if ($company->save()) {
             return redirect()->route('companies.index')
@@ -83,12 +73,13 @@ final class CompaniesController extends Controller
 
 
     /**
-    * Return form to edit existing company.
-    *
-    * @author [Abdullah Alansari] [<ahimta@gmail.com>]
-    * @since [v1.8]
-    * @param int $companyId
-    * @return \Illuminate\Contracts\View\View
+     * Return form to edit existing company.
+     *
+     * @author [Abdullah Alansari] [<ahimta@gmail.com>]
+     * @since [v1.8]
+     * @param int $companyId
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit($companyId)
     {
@@ -107,9 +98,10 @@ final class CompaniesController extends Controller
      *
      * @author [Abdullah Alansari] [<ahimta@gmail.com>]
      * @since [v1.8]
-     * @param Request $request
+     * @param ImageUploadRequest $request
      * @param int $companyId
      * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(ImageUploadRequest $request, $companyId)
     {
@@ -121,37 +113,7 @@ final class CompaniesController extends Controller
 
         $company->name = $request->input('name');
 
-        $old_image = $company->image;
-
-        // Set the model's image property to null if the image is being deleted
-        if ($request->input('image_delete') == 1) {
-            $company->image = null;
-        }
-
-        if ($request->file('image')) {
-            $image = $request->file('image');
-            $file_name = $company->id.'-'.str_slug($image->getClientOriginalName()) . "." . $image->getClientOriginalExtension();
-
-            if ($image->getClientOriginalExtension()!='svg') {
-                Image::make($image->getRealPath())->resize(500, null, function ($constraint) {
-                    $constraint->aspectRatio();
-                    $constraint->upsize();
-                })->save(app('companies_upload_path').$file_name);
-            } else {
-                $image->move(app('companies_upload_path'), $file_name);
-            }
-            $company->image = $file_name;
-
-        }
-
-        if ((($request->file('image')) && (isset($old_image)) && ($old_image!='')) || ($request->input('image_delete') == 1)) {
-            try  {
-                unlink(app('companies_upload_path').$old_image);
-            } catch (\Exception $e) {
-                \Log::error($e);
-            }
-        }
-
+        $company = $request->handleImages($company);
 
         if ($company->save()) {
             return redirect()->route('companies.index')
@@ -162,38 +124,38 @@ final class CompaniesController extends Controller
     }
 
     /**
-    * Delete company
-    *
-    * @author [Abdullah Alansari] [<ahimta@gmail.com>]
-    * @since [v1.8]
-    * @param int $companyId
-    * @return \Illuminate\Http\RedirectResponse
+     * Delete company
+     *
+     * @author [Abdullah Alansari] [<ahimta@gmail.com>]
+     * @since [v1.8]
+     * @param int $companyId
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy($companyId)
     {
+        $this->authorize('delete', $company);
+
         if (is_null($company = Company::find($companyId))) {
             return redirect()->route('companies.index')
                 ->with('error', trans('admin/companies/message.not_found'));
-        } else {
+        }
 
-            $this->authorize('delete', $company);
-
-            try {
-                $company->delete();
+        try {
+            $company->delete();
+            return redirect()->route('companies.index')
+                ->with('success', trans('admin/companies/message.delete.success'));
+        } catch (\Illuminate\Database\QueryException $exception) {
+        /*
+             * NOTE: This happens when there's a foreign key constraint violation
+             * For example when rows in other tables are referencing this company
+             */
+            if ($exception->getCode() == 23000) {
                 return redirect()->route('companies.index')
-                    ->with('success', trans('admin/companies/message.delete.success'));
-            } catch (\Illuminate\Database\QueryException $exception) {
-            /*
-                 * NOTE: This happens when there's a foreign key constraint violation
-                 * For example when rows in other tables are referencing this company
-                 */
-                if ($exception->getCode() == 23000) {
-                    return redirect()->route('companies.index')
-                        ->with('error', trans('admin/companies/message.assoc_users'));
-                } else {
-                    throw $exception;
-                }
+                    ->with('error', trans('admin/companies/message.assoc_users'));
             }
+
+            throw $exception;
         }
     }
 
@@ -203,9 +165,8 @@ final class CompaniesController extends Controller
         if (is_null($company = Company::find($id))) {
             return redirect()->route('companies.index')
                 ->with('error', trans('admin/companies/message.not_found'));
-        } else {
-            return view('companies/view')->with('company',$company);
         }
 
+        return view('companies/view')->with('company',$company);
     }
 }
