+++ app/Http/Controllers/Consumables/ConsumableCheckoutController.php
@@ -1,0 +1,84 @@
+<?php
+
+namespace App\Http\Controllers\Consumables;
+
+use App\Models\Consumable;
+use App\Models\User;
+use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Input;
+
+class ConsumableCheckoutController extends Controller
+{
+
+    /**
+     * Return a view to checkout a consumable to a user.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @see ConsumableCheckoutController::store() method that stores the data.
+     * @since [v1.0]
+     * @param int $consumableId
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
+    public function create($consumableId)
+    {
+        if (is_null($consumable = Consumable::find($consumableId))) {
+            return redirect()->route('consumables.index')->with('error', trans('admin/consumables/message.does_not_exist'));
+        }
+        $this->authorize('checkout', $consumable);
+        return view('consumables/checkout', compact('consumable'));
+    }
+
+    /**
+     * Saves the checkout information
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @see ConsumableCheckoutController::create() method that returns the form.
+     * @since [v1.0]
+     * @param int $consumableId
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
+    public function store($consumableId)
+    {
+        if (is_null($consumable = Consumable::find($consumableId))) {
+            return redirect()->route('consumables.index')->with('error', trans('admin/consumables/message.not_found'));
+        }
+
+        $this->authorize('checkout', $consumable);
+
+        $admin_user = Auth::user();
+        $assigned_to = e(Input::get('assigned_to'));
+
+        // Check if the user exists
+        if (is_null($user = User::find($assigned_to))) {
+            // Redirect to the consumable management page with error
+            return redirect()->route('checkout/consumable', $consumable)->with('error', trans('admin/consumables/message.checkout.user_does_not_exist'));
+        }
+
+        // Update the consumable data
+        $consumable->assigned_to = e(Input::get('assigned_to'));
+
+        $consumable->users()->attach($consumable->id, [
+            'consumable_id' => $consumable->id,
+            'user_id' => $admin_user->id,
+            'assigned_to' => e(Input::get('assigned_to'))
+        ]);
+
+        $logaction = $consumable->logCheckout(e(Input::get('note')), $user);
+        $data['log_id'] = $logaction->id;
+        $data['eula'] = $consumable->getEula();
+        $data['first_name'] = $user->first_name;
+        $data['item_name'] = $consumable->name;
+        $data['checkout_date'] = $logaction->created_at;
+        $data['note'] = $logaction->note;
+        $data['require_acceptance'] = $consumable->requireAcceptance();
+
+
+        // Redirect to the new consumable page
+        return redirect()->route('consumables.index')->with('success', trans('admin/consumables/message.checkout.success'));
+
+    }
+}
