+++ app/Http/Controllers/DepreciationsController.php
@@ -1,15 +1,8 @@
 <?php
 namespace App\Http\Controllers;
 
-use App\Helpers\Helper;
-use Lang;
 use App\Models\Depreciation;
-use Redirect;
-use App\Models\Setting;
-use DB;
-use Str;
-use View;
-use Auth;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 
 /**
@@ -21,13 +14,14 @@ use Illuminate\Http\Request;
 class DepreciationsController extends Controller
 {
     /**
-    * Returns a view that invokes the ajax tables which actually contains
-    * the content for the depreciation listing, which is generated in getDatatable.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net]
-    * @see DepreciationsController::getDatatable() method that generates the JSON response
-    * @since [v1.0]
-    * @return \Illuminate\Contracts\View\View
+     * Returns a view that invokes the ajax tables which actually contains
+     * the content for the depreciation listing, which is generated in getDatatable.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net]
+     * @see DepreciationsController::getDatatable() method that generates the JSON response
+     * @since [v1.0]
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index()
     {
@@ -39,12 +33,13 @@ class DepreciationsController extends Controller
 
 
     /**
-    * Returns a view that displays a form to create a new depreciation.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net]
-    * @see DepreciationsController::postCreate()
-    * @since [v1.0]
-    * @return \Illuminate\Contracts\View\View
+     * Returns a view that displays a form to create a new depreciation.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net]
+     * @see DepreciationsController::postCreate()
+     * @since [v1.0]
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
@@ -63,6 +58,7 @@ class DepreciationsController extends Controller
      * @since [v1.0]
      * @param Request $request
      * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(Request $request)
     {
@@ -84,13 +80,14 @@ class DepreciationsController extends Controller
     }
 
     /**
-    * Returns a view that displays a form to update a depreciation.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net]
-    * @see DepreciationsController::postEdit()
-    * @param int $depreciationId
-    * @since [v1.0]
-    * @return \Illuminate\Contracts\View\View
+     * Returns a view that displays a form to update a depreciation.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net]
+     * @see DepreciationsController::postEdit()
+     * @param int $depreciationId
+     * @since [v1.0]
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit($depreciationId = null)
     {
@@ -115,6 +112,7 @@ class DepreciationsController extends Controller
      * @param int $depreciationId
      * @return \Illuminate\Http\RedirectResponse
      * @since [v1.0]
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(Request $request, $depreciationId = null)
     {
@@ -147,6 +145,7 @@ class DepreciationsController extends Controller
      * @since [v1.0]
      * @param integer $depreciationId
      * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy($depreciationId)
     {
@@ -175,6 +174,7 @@ class DepreciationsController extends Controller
      * @param int $depreciationId
      * @since [v1.0]
      * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show($id)
     {
