+++ app/Http/Controllers/ImportsController.php
@@ -4,17 +4,19 @@ namespace App\Http\Controllers;
 
 use App\Http\Transformers\ImportsTransformer;
 use App\Models\Import;
-use Illuminate\Http\Request;
 use App\Models\Asset;
 
 
 class ImportsController extends Controller
 {
+    /**
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function index()
     {
         $this->authorize('create', Asset::class);
-        $imports = Import::latest()->get();
-        $imports = (new ImportsTransformer)->transformImports($imports);
+        $imports = (new ImportsTransformer)->transformImports(Import::latest()->get());
         return view('importer/import')->with('imports', $imports);
     }
 }
