+++ app/Http/Controllers/LocationsController.php
@@ -1,21 +1,9 @@
 <?php
 namespace App\Http\Controllers;
 
-use App\Helpers\Helper;
-use Input;
-use Lang;
+
 use App\Models\Location;
-use phpDocumentor\Reflection\Types\Array_;
-use Redirect;
-use App\Models\Setting;
-use App\Models\User;
-use App\Models\Asset;
-use DB;
-use Str;
-use Validator;
-use View;
-use Auth;
-use Symfony\Component\HttpFoundation\JsonResponse;
+use Illuminate\Support\Facades\Auth;
 use Image;
 use App\Http\Requests\ImageUploadRequest;
 
@@ -29,13 +17,14 @@ class LocationsController extends Controller
 {
 
     /**
-    * Returns a view that invokes the ajax tables which actually contains
-    * the content for the locations listing, which is generated in getDatatable.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net>]
-    * @see LocationsController::getDatatable() method that generates the JSON response
-    * @since [v1.0]
-    * @return \Illuminate\Contracts\View\View
+     * Returns a view that invokes the ajax tables which actually contains
+     * the content for the locations listing, which is generated in getDatatable.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @see LocationsController::getDatatable() method that generates the JSON response
+     * @since [v1.0]
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index()
     {
@@ -49,12 +38,13 @@ class LocationsController extends Controller
 
 
     /**
-    * Returns a form view used to create a new location.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net>]
-    * @see LocationsController::postCreate() method that validates and stores the data
-    * @since [v1.0]
-    * @return \Illuminate\Contracts\View\View
+     * Returns a form view used to create a new location.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @see LocationsController::postCreate() method that validates and stores the data
+     * @since [v1.0]
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
@@ -72,13 +62,15 @@ class LocationsController extends Controller
 
 
     /**
-    * Validates and stores a new location.
-    *
-    * @todo Check if a Form Request would work better here.
-    * @author [A. Gianotto] [<snipe@snipe.net>]
-    * @see LocationsController::getCreate() method that makes the form
-    * @since [v1.0]
-    * @return \Illuminate\Http\RedirectResponse
+     * Validates and stores a new location.
+     *
+     * @todo Check if a Form Request would work better here.
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @see LocationsController::getCreate() method that makes the form
+     * @since [v1.0]
+     * @param ImageUploadRequest $request
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(ImageUploadRequest $request)
     {
@@ -97,16 +89,7 @@ class LocationsController extends Controller
         $location->manager_id       = $request->input('manager_id');
         $location->user_id          = Auth::id();
 
-        if ($request->file('image')) {
-            $image = $request->file('image');
-            $file_name = str_random(25).".".$image->getClientOriginalExtension();
-            $path = public_path('uploads/locations/'.$file_name);
-            Image::make($image->getRealPath())->resize(600, null, function ($constraint) {
-                $constraint->aspectRatio();
-                $constraint->upsize();
-            })->save($path);
-            $location->image = $file_name;
-        }
+        $location = $request->handleImages($location);
 
         if ($location->save()) {
             return redirect()->route("locations.index")->with('success', trans('admin/locations/message.create.success'));
@@ -116,13 +99,14 @@ class LocationsController extends Controller
 
 
     /**
-    * Makes a form view to edit location information.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net>]
-    * @see LocationsController::postCreate() method that validates and stores
-    * @param int $locationId
-    * @since [v1.0]
-    * @return \Illuminate\Contracts\View\View
+     * Makes a form view to edit location information.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @see LocationsController::postCreate() method that validates and stores
+     * @param int $locationId
+     * @since [v1.0]
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit($locationId = null)
     {
@@ -144,13 +128,15 @@ class LocationsController extends Controller
 
 
     /**
-    * Validates and stores updated location data from edit form.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net>]
-    * @see LocationsController::getEdit() method that makes the form view
-    * @param int $locationId
-    * @since [v1.0]
-    * @return \Illuminate\Http\RedirectResponse
+     * Validates and stores updated location data from edit form.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @see LocationsController::getEdit() method that makes the form view
+     * @param ImageUploadRequest $request
+     * @param int $locationId
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     * @since [v1.0]
      */
     public function update(ImageUploadRequest $request, $locationId = null)
     {
@@ -173,36 +159,7 @@ class LocationsController extends Controller
         $location->ldap_ou      = $request->input('ldap_ou');
         $location->manager_id   = $request->input('manager_id');
 
-        $old_image = $location->image;
-
-        // Set the model's image property to null if the image is being deleted
-        if ($request->input('image_delete') == 1) {
-            $location->image = null;
-        }
-
-        if ($request->file('image')) {
-            $image = $request->file('image');
-            $file_name = $location->id.'-'.str_slug($image->getClientOriginalName()) . "." . $image->getClientOriginalExtension();
-
-            if ($image->getClientOriginalExtension()!='svg') {
-                Image::make($image->getRealPath())->resize(600, null, function ($constraint) {
-                    $constraint->aspectRatio();
-                    $constraint->upsize();
-                })->save(app('locations_upload_path').$file_name);
-            } else {
-                $image->move(app('locations_upload_path'), $file_name);
-            }
-            $location->image = $file_name;
-
-        }
-
-        if ((($request->file('image')) && (isset($old_image)) && ($old_image!='')) || ($request->input('image_delete') == 1)) {
-            try  {
-                unlink(app('locations_upload_path').$old_image);
-            } catch (\Exception $e) {
-                \Log::error($e);
-            }
-        }
+        $location = $request->handleImages($location);
 
 
         if ($location->save()) {
@@ -212,12 +169,13 @@ class LocationsController extends Controller
     }
 
     /**
-    * Validates and deletes selected location.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net>]
-    * @param int $locationId
-    * @since [v1.0]
-    * @return \Illuminate\Http\RedirectResponse
+     * Validates and deletes selected location.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @param int $locationId
+     * @since [v1.0]
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy($locationId)
     {
@@ -226,22 +184,22 @@ class LocationsController extends Controller
             return redirect()->to(route('locations.index'))->with('error', trans('admin/locations/message.not_found'));
         }
 
-        if ($location->users->count() > 0) {
+        if ($location->users()->count() > 0) {
             return redirect()->to(route('locations.index'))->with('error', trans('admin/locations/message.assoc_users'));
 
-        } elseif ($location->childLocations->count() > 0) {
+        } elseif ($location->childLocations()->count() > 0) {
             return redirect()->to(route('locations.index'))->with('error', trans('admin/locations/message.assoc_child_loc'));
 
-        } elseif ($location->assets->count() > 0) {
+        } elseif ($location->assets()->count() > 0) {
             return redirect()->to(route('locations.index'))->with('error', trans('admin/locations/message.assoc_assets'));
 
-        } elseif ($location->assignedassets->count() > 0) {
+        } elseif ($location->assignedassets()->count() > 0) {
             return redirect()->to(route('locations.index'))->with('error', trans('admin/locations/message.assoc_assets'));
 
-        } else {
-            $location->delete();
-            return redirect()->to(route('locations.index'))->with('success', trans('admin/locations/message.delete.success'));
         }
+
+        $location->delete();
+        return redirect()->to(route('locations.index'))->with('success', trans('admin/locations/message.delete.success'));
     }
