+++ app/Http/Controllers/ManufacturersController.php
@@ -1,18 +1,10 @@
 <?php
 namespace App\Http\Controllers;
 
-use App\Helpers\Helper;
 use App\Http\Requests\ImageUploadRequest;
-use App\Models\CustomField;
 use App\Models\Manufacturer;
-use Auth;
-use Exception;
-use Gate;
-use Input;
-use Lang;
+use Illuminate\Support\Facades\Auth;
 use Redirect;
-use Str;
-use View;
 use Illuminate\Http\Request;
 use Image;
 
@@ -25,13 +17,14 @@ use Image;
 class ManufacturersController extends Controller
 {
     /**
-    * Returns a view that invokes the ajax tables which actually contains
-    * the content for the manufacturers listing, which is generated in getDatatable.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net>]
-    * @see Api\ManufacturersController::index() method that generates the JSON response
-    * @since [v1.0]
-    * @return \Illuminate\Contracts\View\View
+     * Returns a view that invokes the ajax tables which actually contains
+     * the content for the manufacturers listing, which is generated in getDatatable.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @see Api\ManufacturersController::index() method that generates the JSON response
+     * @since [v1.0]
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index()
     {
@@ -41,12 +34,13 @@ class ManufacturersController extends Controller
 
 
     /**
-    * Returns a view that displays a form to create a new manufacturer.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net>]
-    * @see ManufacturersController::store()
-    * @since [v1.0]
-    * @return \Illuminate\Contracts\View\View
+     * Returns a view that displays a form to create a new manufacturer.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @see ManufacturersController::store()
+     * @since [v1.0]
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
@@ -61,32 +55,24 @@ class ManufacturersController extends Controller
      * @author [A. Gianotto] [<snipe@snipe.net>]
      * @see ManufacturersController::create()
      * @since [v1.0]
-     * @param Request $request
+     * @param ImageUploadRequest $request
      * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(ImageUploadRequest $request)
     {
 
         $this->authorize('create', Manufacturer::class);
         $manufacturer = new Manufacturer;
-        $manufacturer->name            = $request->input('name');
-        $manufacturer->user_id          = Auth::user()->id;
-        $manufacturer->url     = $request->input('url');
-        $manufacturer->support_url     = $request->input('support_url');
+        $manufacturer->name             = $request->input('name');
+        $manufacturer->user_id          = Auth::id();
+        $manufacturer->url              = $request->input('url');
+        $manufacturer->support_url      = $request->input('support_url');
         $manufacturer->support_phone    = $request->input('support_phone');
         $manufacturer->support_email    = $request->input('support_email');
 
 
-        if ($request->file('image')) {
-            $image = $request->file('image');
-            $file_name = str_slug($image->getClientOriginalName()).".".$image->getClientOriginalExtension();
-            $path = public_path('uploads/manufacturers/'.$file_name);
-            Image::make($image->getRealPath())->resize(200, null, function ($constraint) {
-                $constraint->aspectRatio();
-                $constraint->upsize();
-            })->save($path);
-            $manufacturer->image = $file_name;
-        }
+        $manufacturer = $request->handleImages($manufacturer);
 
 
 
@@ -97,19 +83,20 @@ class ManufacturersController extends Controller
     }
 
     /**
-    * Returns a view that displays a form to edit a manufacturer.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net>]
-    * @see ManufacturersController::update()
-    * @param int $manufacturerId
-    * @since [v1.0]
-    * @return \Illuminate\Contracts\View\View
+     * Returns a view that displays a form to edit a manufacturer.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @see ManufacturersController::update()
+     * @param int $manufacturerId
+     * @since [v1.0]
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function edit($id = null)
+    public function edit($manufacturerId = null)
     {
         $this->authorize('edit', Manufacturer::class);
         // Check if the manufacturer exists
-        if (is_null($item = Manufacturer::find($id))) {
+        if (is_null($item = Manufacturer::find($manufacturerId))) {
             return redirect()->route('manufacturers.index')->with('error', trans('admin/manufacturers/message.does_not_exist'));
         }
         // Show the page
@@ -126,6 +113,7 @@ class ManufacturersController extends Controller
      * @param int $manufacturerId
      * @return \Illuminate\Http\RedirectResponse
      * @since [v1.0]
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(ImageUploadRequest $request, $manufacturerId = null)
     {
@@ -137,42 +125,13 @@ class ManufacturersController extends Controller
         }
 
         // Save the  data
-        $manufacturer->name     = $request->input('name');
-        $manufacturer->url     = $request->input('url');
-        $manufacturer->support_url     = $request->input('support_url');
+        $manufacturer->name             = $request->input('name');
+        $manufacturer->url              = $request->input('url');
+        $manufacturer->support_url      = $request->input('support_url');
         $manufacturer->support_phone    = $request->input('support_phone');
         $manufacturer->support_email    = $request->input('support_email');
 
-        $old_image = $manufacturer->image;
-
-        // Set the model's image property to null if the image is being deleted
-        if ($request->input('image_delete') == 1) {
-            $manufacturer->image = null;
-        }
-
-        if ($request->file('image')) {
-            $image = $request->file('image');
-            $file_name = $manufacturer->id.'-'.str_slug($image->getClientOriginalName()) . "." . $image->getClientOriginalExtension();
-
-            if ($image->getClientOriginalExtension()!='svg') {
-                Image::make($image->getRealPath())->resize(500, null, function ($constraint) {
-                    $constraint->aspectRatio();
-                    $constraint->upsize();
-                })->save(app('manufacturers_upload_path').$file_name);
-            } else {
-                $image->move(app('manufacturers_upload_path'), $file_name);
-            }
-            $manufacturer->image = $file_name;
-
-        }
-
-        if ((($request->file('image')) && (isset($old_image)) && ($old_image!='')) || ($request->input('image_delete') == 1)) {
-            try  {
-                unlink(app('manufacturers_upload_path').$old_image);
-            } catch (\Exception $e) {
-                \Log::error($e);
-            }
-        }
+        $manufacturer = $request->handleImages($manufacturer);
 
 
         if ($manufacturer->save()) {
@@ -182,12 +141,13 @@ class ManufacturersController extends Controller
     }
 
     /**
-    * Deletes a manufacturer.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net>]
-    * @param int $manufacturerId
-    * @since [v1.0]
-    * @return \Illuminate\Http\RedirectResponse
+     * Deletes a manufacturer.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @param int $manufacturerId
+     * @since [v1.0]
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy($manufacturerId)
     {
@@ -219,14 +179,15 @@ class ManufacturersController extends Controller
     }
 
     /**
-    * Returns a view that invokes the ajax tables which actually contains
-    * the content for the manufacturers detail listing, which is generated via API.
-    * This data contains a listing of all assets that belong to that manufacturer.
-    *
-    * @author [A. Gianotto] [<snipe@snipe.net>]
-    * @param int $manufacturerId
-    * @since [v1.0]
-    * @return \Illuminate\Contracts\View\View
+     * Returns a view that invokes the ajax tables which actually contains
+     * the content for the manufacturers detail listing, which is generated via API.
+     * This data contains a listing of all assets that belong to that manufacturer.
+     *
+     * @author [A. Gianotto] [<snipe@snipe.net>]
+     * @param int $manufacturerId
+     * @since [v1.0]
+     * @return \Illuminate\Contracts\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show($manufacturerId = null)
     {
@@ -249,6 +210,7 @@ class ManufacturersController extends Controller
      * @since [v4.1.15]
      * @param int $manufacturers_id
      * @return Redirect
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function restore($manufacturers_id)
     {
