+++ app/Http/Requests/AssetFileRequest.php
@@ -2,7 +2,8 @@
 
 namespace App\Http\Requests;
 
-use App\Http\Requests\Request;
+
+use Symfony\Component\HttpFoundation\JsonResponse;
 
 class AssetFileRequest extends Request
 {
@@ -27,10 +28,5 @@ class AssetFileRequest extends Request
         return [
           'file.*' => 'required|mimes:png,gif,jpg,svg,jpeg,doc,docx,pdf,txt,zip,rar,xls,lic|max:'.$max_file_size,
         ];
-    }
-
-    public function response(array $errors)
-    {
-        return $this->redirector->back()->withInput()->withErrors($errors, $this->errorBag);
     }
 }
