+++ resources/views/users/view.blade.php
@@ -391,7 +391,7 @@
         <div class="tab-pane" id="files_tab">
           <div class="row">
             <div class="col-md-12 col-sm-12">
-              <p>{{ trans('admin/hardware/general.filetype_info') }}</p>
+              <p>{{ trans('admin/users/general.filetype_info') }}</p>
             </div>
             <div class="col-md-2">
               <!-- The fileinput-button span is used to style the file input field as button -->
@@ -427,7 +427,7 @@
 
             <div class="col-md-12 col-sm-12">
               <div class="table-responsive">
-                <table class="display table table-striped">
+                <table id="files-table" class="display table table-striped">
                   <thead>
                     <tr>
                       <th class="col-md-5">{{ trans('general.notes') }}</th>
@@ -560,23 +560,22 @@ $(function () {
 
         done: function (e, data) {
             console.dir(data);
-
             // We use this instead of the fail option, since our API
             // returns a 200 OK status which always shows as "success"
 
-            if (data && data.jqXHR.responseJSON.error && data.jqXHR.responseJSON && data.jqXHR.responseJSON.error) {
-                $('#progress-bar-text').html(data.jqXHR.responseJSON.error);
+            if (data && data.jqXHR && data.jqXHR.responseJSON && data.jqXHR.responseJSON.status === "error") {
+                var errorMessage = data.jqXHR.responseJSON.messages["file.0"];
+                $('#progress-bar-text').html(errorMessage[0]);
                 $('.progress-bar').removeClass('progress-bar-warning').addClass('progress-bar-danger').css('width','100%');
                 $('.progress-checkmark').fadeIn('fast').html('<i class="fa fa-times fa-3x icon-white" style="color: #d9534f"></i>');
-                console.log(data.jqXHR.responseJSON.error);
             } else {
                 $('.progress-bar').removeClass('progress-bar-warning').addClass('progress-bar-success').css('width','100%');
                 $('.progress-checkmark').fadeIn('fast');
                 $('#progress-container').delay(950).css('visibility', 'visible');
                 $('.progress-bar-text').html('Finished!');
                 $('.progress-checkmark').fadeIn('fast').html('<i class="fa fa-check fa-3x icon-white" style="color: green"></i>');
-                $.each(data.result.file, function (index, file) {
-                    $('<tr><td>' + file.notes + '</td><<td>' + file.name + '</td><td>Just now</td><td>' + file.filesize + '</td><td><a class="btn btn-info btn-sm hidden-print" href="import/process/' + file.name + '"><i class="fa fa-spinner process"></i> Process</a></td></tr>').prependTo("#upload-table > tbody");
+                $.each(data.result, function (index, file) {
+                    $('<tr><td>' + file.note + '</td><<td>' + file.filename + '</td></tr>').prependTo("#files-table > tbody");
                 });
             }
             $('#progress').removeClass('active');
