+++ routes/web/consumables.php
@@ -5,15 +5,15 @@
     Route::group([ 'prefix' => 'consumables', 'middleware' => ['auth']], function () {
         Route::get(
             '{consumableID}/checkout',
-            [ 'as' => 'checkout/consumable','uses' => 'ConsumablesController@getCheckout' ]
+            [ 'as' => 'checkout/consumable','uses' => 'Consumables\ConsumableCheckoutController@create' ]
         );
         Route::post(
             '{consumableID}/checkout',
-            [ 'as' => 'checkout/consumable', 'uses' => 'ConsumablesController@postCheckout' ]
+            [ 'as' => 'checkout/consumable', 'uses' => 'Consumables\ConsumableCheckoutController@store' ]
         );
     });
 
-    Route::resource('consumables', 'ConsumablesController', [
+    Route::resource('consumables', 'Consumables\ConsumablesController', [
         'middleware' => ['auth'],
         'parameters' => ['consumable' => 'consumable_id']
     ]);
