+++ routes/web/licenses.php
@@ -4,46 +4,45 @@
 # Licenses
 Route::group([ 'prefix' => 'licenses', 'middleware' => ['auth'] ], function () {
 
-    Route::get('{licenseId}/clone', [ 'as' => 'clone/license', 'uses' => 'LicensesController@getClone' ]);
-    Route::post('{licenseId}/clone', [ 'as' => 'clone/license', 'uses' => 'LicensesController@postCreate' ]);
+    Route::get('{licenseId}/clone', [ 'as' => 'clone/license', 'uses' => 'Licenses\LicensesController@getClone' ]);
 
     Route::get('{licenseId}/freecheckout', [
     'as' => 'licenses.freecheckout',
-    'uses' => 'LicensesController@getFreeLicense'
+    'uses' => 'Licenses\LicensesController@getFreeLicense'
     ]);
     Route::get('{licenseId}/checkout/{seatId?}', [
     'as' => 'licenses.checkout',
-    'uses' => 'LicensesController@getCheckout'
+    'uses' => 'Licenses\LicenseCheckoutController@create'
     ]);
     Route::post(
         '{licenseId}/checkout/{seatId?}',
-        [ 'as' => 'licenses.checkout', 'uses' => 'LicensesController@postCheckout' ]
+        [ 'as' => 'licenses.checkout', 'uses' => 'Licenses\LicenseCheckoutController@store' ]
     );
     Route::get('{licenseId}/checkin/{backto?}', [
     'as' => 'licenses.checkin',
-    'uses' => 'LicensesController@getCheckin'
+    'uses' => 'Licenses\LicenseCheckinController@create'
     ]);
 
     Route::post('{licenseId}/checkin/{backto?}', [
     'as' => 'licenses.checkin.save',
-    'uses' => 'LicensesController@postCheckin'
+    'uses' => 'Licenses\LicenseCheckinController@store'
     ]);
 
     Route::post(
     '{licenseId}/upload',
-    [ 'as' => 'upload/license', 'uses' => 'LicensesController@postUpload' ]
+    [ 'as' => 'upload/license', 'uses' => 'Licenses\LicenseFilesController@store' ]
     );
     Route::delete(
     '{licenseId}/deletefile/{fileId}',
-    [ 'as' => 'delete/licensefile', 'uses' => 'LicensesController@getDeleteFile' ]
+    [ 'as' => 'delete/licensefile', 'uses' => 'Licenses\LicenseFilesController@destroy' ]
     );
     Route::get(
     '{licenseId}/showfile/{fileId}/{download?}',
-    [ 'as' => 'show.licensefile', 'uses' => 'LicensesController@displayFile' ]
+    [ 'as' => 'show.licensefile', 'uses' => 'Licenses\LicenseFilesController@show' ]
     );
 });
 
-Route::resource('licenses', 'LicensesController', [
+Route::resource('licenses', 'Licenses\LicensesController', [
     'middleware' => ['auth'],
     'parameters' => ['license' => 'license_id']
 ]);
